# src/main/java/com/pragma/powerup/application/dto/response/OrderResponseDto.java
package com.pragma.powerup.application.dto.response;

import com.pragma.powerup.domain.enums.OrderStatus;
import lombok.Getter;
import lombok.Setter;
import java.time.Instant;
import java.util.List;

@Getter
@Setter
public class OrderResponseDto {
    private Long id;
    private Long customerId;
    private Long restaurantId;
    private Long assignedEmployeeId;
    private OrderStatus status;
    private Instant createdAt;
    private List<OrderItemResponseDto> items;
}