# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Food Court API
  description: Restaurant management microservice
  version: 1.0.0
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
  /orders:
    get:
      tags: [orders]
      summary: Get restaurant orders filtered by status
      operationId: getOrders
      parameters:
        - in: query
          name: status
          required: true
          schema:
            type: string
            enum: [PENDING, IN_PREPARATION, READY, DELIVERED, CANCELED]
        - in: query
          name: page
          schema: { type: integer, minimum: 0, default: 0 }
        - in: query
          name: size
          schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
      responses:
        '200':
          description: Orders from the authenticated employee's restaurant
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderPage' }
        '400': { description: Invalid status or pagination }
        '401': { description: Authentication required }
        '403': { description: EMPLOYEE role required }
    post:
      tags: [orders]
      summary: Create an order
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OrderRequest' }
      responses:
        '201':
          description: Order created with PENDING status
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderResponseEnvelope' }
        '400': { description: Invalid dishes, quantities or customer already has an active order }
        '401': { description: Authentication required }
        '403': { description: CUSTOMER role required }
        '404': { description: Restaurant not found }
        '503': { description: Traceability service unavailable }
  /restaurants:
    get:
      tags: [restaurants]
      summary: Get restaurants alphabetically
      operationId: getRestaurants
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 0, default: 0 }
        - in: query
          name: size
          schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
      responses:
        '200':
          description: Paginated restaurant summaries ordered by name
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RestaurantSummaryPage' }
        '400': { description: Invalid pagination parameters }
        '401': { description: Authentication required }
        '403': { description: CUSTOMER role required }
    post:
      tags: [restaurants]
      summary: Create a restaurant
      operationId: createRestaurant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestaurantRequest'
      responses:
        '201':
          description: Restaurant created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestaurantResponseEnvelope'
        '400': { description: Invalid fields, duplicated NIT or owner without OWNER role }
        '503': { description: User service unavailable }
  /restaurants/{restaurantId}/ownership:
    get:
      tags: [restaurants]
      summary: Validate that the authenticated owner owns a restaurant
      operationId: validateOwnership
      parameters:
        - in: path
          name: restaurantId
          required: true
          schema: { type: integer, format: int64, minimum: 1 }
      responses:
        '204': { description: Restaurant belongs to the authenticated owner }
        '401': { description: Authentication required }
        '403': { description: OWNER role or ownership required }
        '404': { description: Restaurant not found }
  /restaurants/{restaurantId}/dishes:
    get:
      tags: [dishes]
      summary: Get active dishes from a restaurant
      operationId: getDishes
      parameters:
        - in: path
          name: restaurantId
          required: true
          schema: { type: integer, format: int64, minimum: 1 }
        - in: query
          name: categoryId
          required: false
          schema: { type: integer, format: int64, minimum: 1 }
        - in: query
          name: page
          schema: { type: integer, minimum: 0, default: 0 }
        - in: query
          name: size
          schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
      responses:
        '200':
          description: Paginated active dishes, optionally filtered by category
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DishPage' }
        '400': { description: Invalid pagination or category parameters }
        '401': { description: Authentication required }
        '403': { description: CUSTOMER role required }
        '404': { description: Restaurant not found }
    post:
      tags: [dishes]
      summary: Create a dish for a restaurant
      operationId: createDish
      parameters:
        - in: path
          name: restaurantId
          required: true
          schema: { type: integer, format: int64, minimum: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DishRequest' }
      responses:
        '201':
          description: Dish created and enabled
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DishResponseEnvelope' }
        '400': { description: Invalid fields, restaurant or category }
  /restaurants/{restaurantId}/dishes/{dishId}:
    patch:
      tags: [dishes]
      summary: Update a dish's price and description
      operationId: updateDish
      parameters:
        - in: path
          name: restaurantId
          required: true
          schema: { type: integer, format: int64, minimum: 1 }
        - in: path
          name: dishId
          required: true
          schema: { type: integer, format: int64, minimum: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DishUpdateRequest' }
      responses:
        '200':
          description: Dish updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DishResponseEnvelope' }
        '400': { description: Invalid price or description }
        '404': { description: Dish does not exist in the restaurant }
  /restaurants/{restaurantId}/dishes/{dishId}/status:
    patch:
      tags: [dishes]
      summary: Enable or disable a dish
      operationId: updateDishStatus
      parameters:
        - in: path
          name: restaurantId
          required: true
          schema: { type: integer, format: int64, minimum: 1 }
        - in: path
          name: dishId
          required: true
          schema: { type: integer, format: int64, minimum: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DishStatusRequest' }
      responses:
        '200':
          description: Dish status updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DishResponseEnvelope' }
        '400': { description: Status is required }
        '401': { description: Authentication required }
        '403': { description: Only the restaurant owner can update the dish }
        '404': { description: Dish does not exist in the restaurant }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RestaurantRequest:
      type: object
      required: [name, nit, address, phone, urlLogo, ownerId]
      properties:
        name:
          type: string
          pattern: '.*[^0-9].*'
          example: El Corral 2
        nit:
          type: string
          pattern: '^\d+$'
          example: '900123456'
        address: { type: string, example: 'Local 15' }
        phone:
          type: string
          maxLength: 13
          pattern: '^\+?\d+$'
          example: '+573005698325'
        urlLogo: { type: string, example: 'https://cdn.example.com/logo.png' }
        ownerId: { type: integer, format: int64, minimum: 1 }
    RestaurantResponse:
      allOf:
        - $ref: '#/components/schemas/RestaurantRequest'
        - type: object
          properties:
            id: { type: integer, format: int64 }
    RestaurantResponseEnvelope:
      type: object
      properties:
        data: { $ref: '#/components/schemas/RestaurantResponse' }
    RestaurantSummary:
      type: object
      properties:
        name: { type: string, example: El Corral }
        urlLogo: { type: string, example: 'https://cdn.example.com/logo.png' }
    PageMetadata:
      type: object
      properties:
        page: { type: integer, example: 0 }
        size: { type: integer, example: 10 }
        totalElements: { type: integer, format: int64, example: 25 }
        totalPages: { type: integer, example: 3 }
    RestaurantSummaryPage:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/RestaurantSummary' }
        meta: { $ref: '#/components/schemas/PageMetadata' }
    DishRequest:
      type: object
      required: [name, price, description, urlImage, categoryId]
      properties:
        name: { type: string, example: Hamburguesa especial }
        price: { type: integer, format: int64, minimum: 1, example: 25000 }
        description: { type: string, example: Carne, queso y vegetales }
        urlImage: { type: string, example: 'https://cdn.example.com/dish.png' }
        categoryId: { type: integer, format: int64, minimum: 1, example: 2 }
    DishResponse:
      allOf:
        - $ref: '#/components/schemas/DishRequest'
        - type: object
          properties:
            id: { type: integer, format: int64 }
            restaurantId: { type: integer, format: int64 }
            active: { type: boolean, example: true }
    DishUpdateRequest:
      type: object
      required: [price, description]
      properties:
        price: { type: integer, format: int64, minimum: 1, example: 30000 }
        description: { type: string, example: Nueva descripción del plato }
    DishStatusRequest:
      type: object
      required: [active]
      properties:
        active: { type: boolean, example: false }
    DishResponseEnvelope:
      type: object
      properties:
        data: { $ref: '#/components/schemas/DishResponse' }
    DishPage:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/DishResponse' }
        meta: { $ref: '#/components/schemas/PageMetadata' }
    OrderItemRequest:
      type: object
      required: [dishId, quantity]
      properties:
        dishId: { type: integer, format: int64, minimum: 1 }
        quantity: { type: integer, minimum: 1 }
    OrderRequest:
      type: object
      required: [restaurantId, items]
      properties:
        restaurantId: { type: integer, format: int64, minimum: 1 }
        items:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/OrderItemRequest' }
    OrderResponse:
      allOf:
        - $ref: '#/components/schemas/OrderRequest'
        - type: object
          properties:
            id: { type: integer, format: int64 }
            customerId: { type: integer, format: int64 }
            assignedEmployeeId: { type: integer, format: int64, nullable: true }
            status: { type: string, enum: [PENDING, IN_PREPARATION, READY, DELIVERED, CANCELED] }
            createdAt: { type: string, format: date-time }
    OrderResponseEnvelope:
      type: object
      properties:
        data: { $ref: '#/components/schemas/OrderResponse' }
    OrderPage:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/OrderResponse' }
        meta: { $ref: '#/components/schemas/PageMetadata' }